Improve security by hashing tokens in database
It's not the most secure it could possibly be, but it's pretty good. If the hash was salted, it would be more secure, but super slow to use. If the database gets leaked, the tokens will all be reset anyway. The idea of doing this came about due to the news of the recent leak of twitch.tv.